moved PlayerHider into modules.playerhider package
